Made NLopt.jl interface properly handle options & minor changes to MOI
NLopt.jl interface now takes option for global optimiser as kwargs and if a `local_method` is defined  maxiters and option for the local optimiser can be passed as `local_maxiters` and a NamedTuple to `local_options` respectively.

Moreover, the interface now takes both `NLopt.Opt` and the newly defined `NLO` type as optimiser input. The `NLO` type follows the same logic of `BBO`. 

Also I removed n_starts argument as it was passed to the optimiser and only used to multiply the `maxiters`. 

The NLopt `__solve` also passes `ret` to the `retcode` of `SciMLBase.SciMLBase.OptimizationSolution` and the NLopt model is passed to `original` field.

Additionally, for `MOI` interface now returns the `MOI.TerminationStatus()` via the `retcode` field and the optimised `MOI` problem via the `original` field of `SciMLBase.SciMLBase.OptimizationSolution`.

I have added some test for the changed interface