Remove "higher level docs" and put a higher level table
It would be difficult to have 2 versions of the documentation for the methods, one done by type (local vs global, etc.) and one by package. It would be nice in the future to generate that from tabular data, but for now let's just stick to the per-package documentation and put a table on the front page to help people figure out which packages may be useful.