Makes GalacticOptim load MultistartOptimization depend on version
introduces a function which checks the version and if master is branch is checked out of loaded optimizer package. This might useful more general since there is no Base Julia way of doing this at the moment for "soft" dependencies.

Amended the docs to reflect the changed interface if user have `MultistartOptimization` v0.1.2#master or higher installed.

Unfortunately, I can't create a test case for this since from what I understand I can't force the test environment to install from the master branch of `MultistartOptimization`.