Reverted back to a single file without if statement
Changed back to a single file with `@require` based  loading of the original `__solve` function which dispatches on `NLopt.Algotrithm` as the `local_method`. 

The new `__solve` function which calls the new generic API of `MultistartOptimization` has three positional argument and allows the usage of all `GalacticOptim` optimizers. This will simply error now if the `MultistartOptimization` is not up to date enough since the underlying API doesn't exist yet